refactor: rename utils to wasm-interop
